+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - add set token in single
 
+- new component pie chart
+
 ## [1.44.26] [newSetting]
 ### Adde